+++ CHANGELOG.md
@@ -1,4 +1,2 @@
 - Experimental feature: qst will now use an aethernet shortcut as a hint for where to teleport to for the automatic teleporting thing. I've tested it a bit, but it probably still has issues, please let me know how that goes. -alydev
-- Feature: /qst zone now takes an ID argument -alydev
-- Better pathfinding for warring triad unlocks -Kiarra
-- landing fixes for 3673 -Kiarra
+- Feature: Notice and command to clear completion log -alydev